requirements.txt
python-telegram-bot
alpha_vantage
pytz
pandas
pandas_ta